requirements.txt
requests
datetime
flask
discord.py
python-dotenv